# docs/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
      "ApiError" : {
        "description" : "Error response body.",
        "type" : "object",
        "required" : [ "error", "message" ],
        "properties" : {
          "error" : {
            "type" : "string",
            "description" : "Machine-readable error code."
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable description of the error."
          }
        }
      },
      "BatchEvaluationRequest" : {
        "description" : "Batch of authorization evaluation requests.",
        "type" : "object",
        "required" : [ "requests" ],
        "properties" : {
          "requests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EvaluationRequest"
            },
            "description" : "Requests to evaluate. Results match the input order."
          }
        }
      },
      "BatchEvaluationResult" : {
        "description" : "Batch evaluation result.",
        "type" : "object",
        "required" : [ "decisions" ],
        "properties" : {
          "decisions" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Decision"
            },
            "description" : "One decision per input request, in the same order."
          }
        }
      },
      "Decision" : {
        "type" : "object",
        "description" : "Authorization decision from the PDP.",
        "required" : [ "allowed", "reason", "decisionId", "policyVersion" ],
        "properties" : {
          "allowed" : {
            "type" : "boolean",
            "description" : "true if the action is permitted."
          },
          "reason" : {
            "type" : "string",
            "description" : "Human-readable explanation for the decision."
          },
          "decisionId" : {
            "type" : "string",
            "description" : "Unique identifier for this decision instance (UUID)."
          },
          "policyVersion" : {
            "type" : "string",
            "description" : "Policy version that produced the decision."
          },
          "obligations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Obligation"
            },
            "description" : "Obligations the PEP must fulfill when enforcing this decision."
          }
        }
      },
      "EvaluationRequest" : {
        "type" : "object",
        "description" : "Authorization evaluation request from a PEP.",
        "required" : [ "action", "resource" ],
        "properties" : {
          "action" : {
            "type" : "string",
            "description" : "Action in 'resource:verb' format, e.g. 'empleado:read'."
          },
          "resource" : {
            "description" : "Target resource being accessed.",
            "type" : "object",
            "$ref" : "#/components/schemas/ResourceRef"
          },
          "context" : {
            "type" : "object",
            "additionalProperties" : { },
            "description" : "Optional runtime context attributes, e.g. {\"emergency\": true}."
          }
        }
      },
      "Obligation" : {
        "type" : "object",
        "description" : "Obligation the PEP must fulfill when enforcing a permit or deny decision.",
        "required" : [ "type" ],
        "properties" : {
          "type" : {
            "type" : "string",
            "description" : "Obligation type identifier."
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : { },
            "description" : "Obligation-specific parameters."
          }
        }
      },
      "PermissionsResponse" : {
        "description" : "Cacheable flat list of permitted actions for the authenticated subject.",
        "type" : "object",
        "required" : [ "subject", "app", "permissions", "policyVersion", "evaluatedAt" ],
        "properties" : {
          "subject" : {
            "type" : "string",
            "description" : "Resolved subject identifier from the JWT."
          },
          "app" : {
            "type" : "string",
            "description" : "Application context for which permissions are listed."
          },
          "permissions" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            },
            "description" : "Permitted actions in 'resource:verb' format."
          },
          "policyVersion" : {
            "type" : "string",
            "description" : "Policy version that produced this response."
          },
          "evaluatedAt" : {
            "type" : "string",
            "description" : "ISO-8601 instant when this response was generated."
          }
        }
      },
      "ResourceRef" : {
        "type" : "object",
        "description" : "Target resource being accessed.",
        "required" : [ "type" ],
        "properties" : {
          "type" : {
            "type" : "string",
            "description" : "Resource type, e.g. 'empleado'."
          },
          "id" : {
            "type" : "string",
            "description" : "Unique resource identifier. Null for collection-level checks."
          },
          "attributes" : {
            "type" : "object",
            "additionalProperties" : { },
            "description" : "Optional inline resource attributes, keyed by attribute id or code."
          }
        }
      }
    }
  },
  "tags" : [ {
    "name" : "Metadata",
    "description" : "Service metadata and information"
  }, {
    "name" : "evaluation",
    "description" : "PEP-facing authorization evaluation endpoints."
  }, {
    "name" : "permissions",
    "description" : "Subject permission listing endpoint."
  } ],
  "paths" : {
    "/info" : {
      "get" : {
        "summary" : "Get service information",
        "description" : "Returns metadata about this Service Policy instance, including name, version, repository URL, and uptime.",
        "tags" : [ "Metadata" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : { }
                }
              }
            }
          }
        }
      }
    },
    "/v1/evaluate" : {
      "post" : {
        "summary" : "Evaluate a single authorization request",
        "description" : "Evaluates whether the authenticated subject may perform the requested action on the given resource. Subject is resolved from the Bearer JWT in the Authorization header. Returns HTTP 401 if the header is missing or malformed. Returns HTTP 400 if 'action' or 'resource.type' is blank.",
        "tags" : [ "evaluation" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/v1/evaluate/batch" : {
      "post" : {
        "summary" : "Evaluate a batch of authorization requests",
        "description" : "Evaluates multiple requests in a single call. Results are returned in the same order as the input requests. Requires a valid Bearer token in the Authorization header.",
        "tags" : [ "evaluation" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BatchEvaluationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        }
      }
    },
    "/v1/permissions" : {
      "get" : {
        "summary" : "List permitted actions for the authenticated subject",
        "description" : "Returns a flat, cacheable list of actions the subject is permitted to perform within the given application context. Subject is resolved from the Bearer JWT in the Authorization header. The response is safe to cache per subject + app + policyVersion; the policyVersion field changes when policies are updated. Returns HTTP 401 if the Authorization header is missing or malformed. Returns HTTP 400 if the 'app' query parameter is missing or blank.",
        "tags" : [ "permissions" ],
        "parameters" : [ {
          "name" : "app",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        }
      }
    }
  },
  "info" : {
    "title" : "Service Policy — PDP",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}